+++ driver/src/main/java/oracle/nosql/driver/iam/FederationRequestHelper.java
@@ -10,6 +10,7 @@ package oracle.nosql.driver.iam;
 import static oracle.nosql.driver.iam.Utils.*;
 import static oracle.nosql.driver.util.HttpConstants.*;
 import static oracle.nosql.driver.util.LogUtil.logHeaderValue;
+import static oracle.nosql.driver.util.LogUtil.redactSensitiveValue;
 
 import java.io.IOException;
 import java.io.StringWriter;
@@ -74,9 +75,10 @@ class FederationRequestHelper {
                     "Error getting security token from IAM, " +
                     "status code %d, response \n%s",
                     response.getStatusCode(),
-                    response.getOutput()));
+                    redactSensitiveValue(response.getOutput())));
         }
-        logTrace(logger, "Federation response " + response.getOutput());
+        logTrace(logger, "Federation response " +
+                 redactSensitiveValue(response.getOutput()));
         return parseTokenResponse(response.getOutput());
     }